Introducing RenderFn
This commit adds the new RenderFn type which all renders now
are based on.

The signature is:
RenderFn<DataProps, Config>: JSX.Element

Alongside, to maintain compatibility with previous renders, we
introduce the RenderFnWithoutData, which does not receive any data
prop. It's the case of delimiter renderer.

No existing tests were touch to maintain backward compatibility.